feat: rework organization wallet modal
Replace the single-view wallet overlay with a Balances/Manage tab split and
break the monolithic component into focused modules under components/overlays/wallet/.

Balances tab:
- Ethereum, Base, and Tempo are surfaced first in both Mainnets and Testnets
  (via CHAIN_DISPLAY_ORDER in chain-utils)
- Each chain is folded by default and auto-expands once balance loading
  resolves and the chain has a positive native, stablecoin, or custom token
  balance. First-paint skips the Radix slide animation so switching filters
  does not feel laggy
- Per-chain inline "Add token" input replaces the global form. The chain is
  implicit, and 42-char input is EIP-55 checksummed on change and on submit

Manage tab:
- Wallet address card with click-to-copy affordance (3s green confirmation
  state, matching copy icon swap)
- Recovery email card with inline edit flow
- Security card gates Export Private Key behind admin + Turnkey wallet
- Cards drop the muted background to match other overlays (API keys, etc.)

Tab state is controlled so reload cycles (switching Para/Turnkey, refresh)
do not throw the user back to Balances.
