+++ components/overlays/wallet-overlay.tsx
@@ -1,1178 +1,39 @@
 "use client";
 
-import {
-  Copy,
-  Eye,
-  EyeOff,
-  ExternalLink,
-  KeyRound,
-  Plus,
-  RefreshCw,
-  SendHorizontal,
-  ShieldCheck,
-  Trash2,
-} from "lucide-react";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import Image from "next/image";
 import { useCallback, useEffect, useState } from "react";
 import { toast } from "sonner";
 import { Overlay } from "@/components/overlays/overlay";
 import { useOverlay } from "@/components/overlays/overlay-provider";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { Spinner } from "@/components/ui/spinner";
-import { toChecksumAddress, truncateAddress } from "@/lib/address-utils";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { useSession } from "@/lib/auth-client";
 import { useActiveMember } from "@/lib/hooks/use-organization";
 import { fetchAllSupportedTokenBalances } from "@/lib/wallet/fetch-balances";
 import type {
-  ChainBalance,
   ChainData,
   SupportedToken,
   SupportedTokenBalance,
-  TokenBalance,
   TokenData,
   WalletData,
-  WalletInfo,
 } from "@/lib/wallet/types";
 import { useWalletBalances } from "@/lib/wallet/use-wallet-balances";
+import { BalancesTab } from "./wallet/balances-tab";
+import { isTempoChain } from "./wallet/chain-utils";
+import { ManageTab } from "./wallet/manage-tab";
+import { NoWalletSection } from "./wallet/no-wallet-section";
 import { type WithdrawableAsset, WithdrawModal } from "./withdraw-modal";
+
+type WalletTab = "balances" | "manage";
 
 type WalletOverlayProps = {
   overlayId: string;
+  initialTab?: WalletTab;
 };
 
-// TEMPO uses stablecoins for gas, so we display stablecoins only (no native token)
-const TEMPO_CHAIN_IDS = new Set([42_429, 4217]);
-const isTempoChain = (chainId: number): boolean => TEMPO_CHAIN_IDS.has(chainId);
-
-// Chains whose token lineup doesn't mirror Ethereum mainnet's stablecoin set
-// (e.g. Plasma ships USDT0, no Circle USDC, no Sky USDS). For these chains we
-// render the chain's own supported_tokens rows directly instead of overlaying
-// them on the mainnet master list, which would otherwise produce misleading
-// "Not available" entries for assets that simply don't exist on the chain.
-const INDEPENDENT_TOKEN_LIST_CHAIN_IDS = new Set([42_429, 4217, 9745]);
-const hasIndependentTokenList = (chainId: number): boolean =>
-  INDEPENDENT_TOKEN_LIST_CHAIN_IDS.has(chainId);
-
-const MAINNET_CHAIN_ID = 1;
-
-// ============================================================================
-// Balance Display Components
-// ============================================================================
-
-function ChainBalanceDisplay({ balance }: { balance: ChainBalance }) {
-  if (balance.loading) {
-    return <div className="mt-1 text-muted-foreground text-xs">Loading...</div>;
-  }
-  if (balance.error) {
-    return <div className="mt-1 text-destructive text-xs">{balance.error}</div>;
-  }
-  return (
-    <div className="mt-1 text-muted-foreground text-xs">
-      {balance.balance} {balance.symbol}
-    </div>
-  );
-}
-
-// Token item with withdraw and optional delete button
-function TokenItemWithActions({
-  token,
-  isAdmin,
-  isCustom,
-  customExplorerUrl,
-  onDelete,
-  onWithdraw,
-}: {
-  token: SupportedTokenBalance | TokenBalance;
-  isAdmin: boolean;
-  isCustom?: boolean;
-  customExplorerUrl?: string | null;
-  onDelete?: (tokenId: string, symbol: string) => void;
-  onWithdraw: (chainId: number, tokenAddress: string) => void;
-}) {
-  // Type guard to check if it's a SupportedTokenBalance (has logoUrl)
-  const isSupportedToken = (
-    t: SupportedTokenBalance | TokenBalance
-  ): t is SupportedTokenBalance => "logoUrl" in t;
-
-  const supportedToken = isSupportedToken(token) ? token : null;
-  const customToken = isSupportedToken(token) ? null : token;
-
-  const isUnavailable = supportedToken?.available === false;
-  const numBalance = Number.parseFloat(token.balance);
-  const hasBalance = Number.isFinite(numBalance) && numBalance > 0;
-
-  // Both types now have tokenAddress
-  const tokenAddress =
-    supportedToken?.tokenAddress || customToken?.tokenAddress;
-
-  // Explorer URL: use supported token's URL or custom explorer URL for custom tokens
-  const explorerUrl = supportedToken?.explorerUrl || customExplorerUrl;
-
-  const copyTokenAddress = () => {
-    if (tokenAddress) {
-      navigator.clipboard.writeText(tokenAddress);
-      toast.success("Token address copied");
-    }
-  };
-
-  const renderBalance = () => {
-    if (isUnavailable) {
-      return <span className="italic">Not available</span>;
-    }
-    if (token.loading) {
-      return <Spinner className="h-3 w-3" />;
-    }
-    if (token.error) {
-      return <span className="text-destructive">{token.error}</span>;
-    }
-    return `${token.balance} ${token.symbol}`;
-  };
-
-  return (
-    <div
-      className={`flex items-center gap-2 py-1.5 ${isUnavailable ? "opacity-50" : ""}`}
-    >
-      {/* Logo: supported tokens show logo, custom tokens show $ */}
-      {supportedToken?.logoUrl && (
-        <Image
-          alt={token.symbol}
-          className={`h-4 w-4 rounded-full ${isUnavailable ? "grayscale" : ""}`}
-          height={16}
-          src={supportedToken.logoUrl}
-          width={16}
-        />
-      )}
-      {!supportedToken?.logoUrl && (
-        <div className="flex h-4 w-4 items-center justify-center rounded-full bg-muted font-medium text-[10px]">
-          $
-        </div>
-      )}
-      <span className="font-medium text-xs">{token.symbol}</span>
-      {/* Copy + explorer link for all tokens with addresses */}
-      {!isUnavailable && tokenAddress && (
-        <div className="flex items-center gap-1">
-          <button
-            aria-label="Copy token address"
-            className="text-muted-foreground hover:text-foreground"
-            onClick={copyTokenAddress}
-            type="button"
-          >
-            <Copy className="h-3 w-3" />
-          </button>
-          {explorerUrl && (
-            <a
-              className="text-muted-foreground hover:text-foreground"
-              href={explorerUrl}
-              rel="noopener noreferrer"
-              target="_blank"
-              title="View on explorer"
-            >
-              <ExternalLink className="h-3 w-3" />
-            </a>
-          )}
-        </div>
-      )}
-      {/* Delete button for custom tokens - before balance */}
-      {isAdmin && isCustom && customToken && onDelete && (
-        <Button
-          className="ml-auto h-6 w-6"
-          onClick={() => onDelete(customToken.tokenId, customToken.symbol)}
-          size="icon"
-          title="Remove token"
-          variant="ghost"
-        >
-          <Trash2 className="h-3 w-3 text-muted-foreground" />
-        </Button>
-      )}
-      {/* Balance */}
-      <span
-        className={`text-muted-foreground text-xs ${isAdmin && isCustom ? "" : "ml-auto"}`}
-      >
-        {renderBalance()}
-      </span>
-      {/* Withdraw button for tokens with balance */}
-      {isAdmin &&
-        hasBalance &&
-        !token.loading &&
-        !isUnavailable &&
-        tokenAddress && (
-          <Button
-            className="h-6 px-2 text-xs"
-            onClick={() => onWithdraw(token.chainId, tokenAddress)}
-            size="sm"
-            variant="ghost"
-          >
-            <SendHorizontal className="h-3 w-3" />
-          </Button>
-        )}
-    </div>
-  );
-}
-
-// Helper to build explorer URL for a token address
-function buildTokenExplorerUrl(
-  chain: ChainData | undefined,
-  tokenAddress: string
-): string | null {
-  if (!chain?.explorerUrl) {
-    return null;
-  }
-  const path = chain.explorerAddressPath || "/address/{address}";
-  return `${chain.explorerUrl}${path.replace("{address}", tokenAddress)}`;
-}
-
-function ChainBalanceItem({
-  balance,
-  chain,
-  isAdmin,
-  onRemoveToken,
-  onWithdraw,
-  supportedTokenBalances,
-  tokenBalances,
-}: {
-  balance: ChainBalance;
-  chain: ChainData | undefined;
-  isAdmin: boolean;
-  onRemoveToken: (tokenId: string, symbol: string) => void;
-  onWithdraw: (chainId: number, tokenAddress?: string) => void;
-  supportedTokenBalances: SupportedTokenBalance[];
-  tokenBalances: TokenBalance[];
-}) {
-  const chainCustomTokens = tokenBalances.filter(
-    (t) => t.chainId === balance.chainId
-  );
-
-  const isTempo = isTempoChain(balance.chainId);
-  const isMainnet = balance.chainId === MAINNET_CHAIN_ID;
-  const isIndependentTokenList = hasIndependentTokenList(balance.chainId);
-
-  // For chains with their own stablecoin lineup (TEMPO, Plasma, etc.), render
-  // their tokens directly. For other chains, use mainnet tokens as the master
-  // list and overlay availability per-chain.
-  const chainSupportedTokens = (() => {
-    if (isIndependentTokenList) {
-      return supportedTokenBalances.filter(
-        (t) => t.chainId === balance.chainId
-      );
-    }
-
-    // Get mainnet tokens as master list
-    const mainnetTokens = supportedTokenBalances.filter(
-      (t) => t.chainId === MAINNET_CHAIN_ID
-    );
-
-    // If viewing mainnet, just return mainnet tokens
-    if (isMainnet) {
-      return mainnetTokens;
-    }
-
-    // For other chains, map mainnet tokens with availability
-    const chainTokensMap = new Map(
-      supportedTokenBalances
-        .filter((t) => t.chainId === balance.chainId)
-        .map((t) => [t.symbol, t])
-    );
-
-    return mainnetTokens.map((mainnetToken) => {
-      const chainToken = chainTokensMap.get(mainnetToken.symbol);
-      if (chainToken) {
-        // Token available on this chain
-        return { ...chainToken, available: true };
-      }
-      // Token not available - show mainnet data with unavailable flag
-      return { ...mainnetToken, available: false, balance: "N/A" };
-    });
-  })();
-
-  // For non-TEMPO chains, check if native balance is withdrawable
-  const nativeBalance = Number.parseFloat(balance.balance);
-  const hasNativeBalance = Number.isFinite(nativeBalance) && nativeBalance > 0;
-
-  // Determine section label based on whether there are custom tokens
-  const tokenSectionLabel =
-    chainCustomTokens.length > 0 ? "Tokens" : "Stablecoins";
-
-  // For TEMPO, show tokens only (no native balance display)
-  if (isTempo) {
-    return (
-      <div className="rounded-lg border bg-muted/50 p-3">
-        <div className="mb-2 flex items-center gap-1">
-          <span className="font-medium text-sm">TEMPO</span>
-          {balance.explorerUrl && (
-            <a
-              className="text-muted-foreground hover:text-foreground"
-              href={balance.explorerUrl}
-              rel="noopener noreferrer"
-              target="_blank"
-            >
-              <ExternalLink className="h-3 w-3" />
-            </a>
-          )}
-        </div>
-        <div className="mb-1 font-medium text-muted-foreground text-xs">
-          {tokenSectionLabel}
-        </div>
-        {chainSupportedTokens.length > 0 || chainCustomTokens.length > 0 ? (
-          <div className="divide-y rounded border bg-background/50 px-2">
-            {/* Supported tokens (stablecoins) */}
-            {chainSupportedTokens.map((token) => (
-              <TokenItemWithActions
-                isAdmin={isAdmin}
-                key={`supported-${token.chainId}-${token.tokenAddress}`}
-                onWithdraw={onWithdraw}
-                token={token}
-              />
-            ))}
-            {/* Custom tokens integrated in the same list */}
-            {chainCustomTokens.map((token) => (
-              <TokenItemWithActions
-                customExplorerUrl={buildTokenExplorerUrl(
-                  chain,
-                  token.tokenAddress
-                )}
-                isAdmin={isAdmin}
-                isCustom
-                key={`custom-${token.tokenId}`}
-                onDelete={onRemoveToken}
-                onWithdraw={onWithdraw}
-                token={token}
-              />
-            ))}
-          </div>
-        ) : (
-          <div className="text-muted-foreground text-xs">Loading tokens...</div>
-        )}
-      </div>
-    );
-  }
-
-  // Non-TEMPO chains: show native balance and tokens in same card
-  return (
-    <div className="rounded-lg border bg-muted/50 p-3">
-      {/* Chain header with native balance */}
-      <div className="flex items-center justify-between">
-        <div className="flex-1">
-          <div className="flex items-center gap-1">
-            <span className="font-medium text-sm">{balance.name}</span>
-            {balance.explorerUrl && (
-              <a
-                className="text-muted-foreground hover:text-foreground"
-                href={balance.explorerUrl}
-                rel="noopener noreferrer"
-                target="_blank"
-              >
-                <ExternalLink className="h-3 w-3" />
-              </a>
-            )}
-          </div>
-          <ChainBalanceDisplay balance={balance} />
-        </div>
-        {isAdmin && hasNativeBalance && (
-          <Button
-            className="h-7 px-2 text-xs"
-            onClick={() => onWithdraw(balance.chainId)}
-            size="sm"
-            variant="ghost"
-          >
-            <SendHorizontal className="h-3 w-3" />
-            Withdraw
-          </Button>
-        )}
-      </div>
-      {/* Tokens section (stablecoins + custom tokens combined) */}
-      {(chainSupportedTokens.length > 0 || chainCustomTokens.length > 0) && (
-        <div className="mt-3">
-          <div className="mb-1 font-medium text-muted-foreground text-xs">
-            {tokenSectionLabel}
-          </div>
-          <div className="divide-y rounded border bg-background/50 px-2">
-            {/* Supported tokens (stablecoins) */}
-            {chainSupportedTokens.map((token) => (
-              <TokenItemWithActions
-                isAdmin={isAdmin}
-                key={`supported-${token.chainId}-${token.tokenAddress}`}
-                onWithdraw={onWithdraw}
-                token={token}
-              />
-            ))}
-            {/* Custom tokens integrated in the same list */}
-            {chainCustomTokens.map((token) => (
-              <TokenItemWithActions
-                customExplorerUrl={buildTokenExplorerUrl(
-                  chain,
-                  token.tokenAddress
-                )}
-                isAdmin={isAdmin}
-                isCustom
-                key={`custom-${token.tokenId}`}
-                onDelete={onRemoveToken}
-                onWithdraw={onWithdraw}
-                token={token}
-              />
-            ))}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ============================================================================
-// Form Components
-// ============================================================================
-
-function AddTokenForm({
-  chains,
-  onAdd,
-  onCancel,
-}: {
-  chains: ChainData[];
-  onAdd: (chainId: number, tokenAddress: string) => Promise<void>;
-  onCancel: () => void;
-}) {
-  const [chainId, setChainId] = useState<number | null>(null);
-  const [tokenAddress, setTokenAddress] = useState("");
-  const [adding, setAdding] = useState(false);
-
-  const handleAdd = async () => {
-    if (!(chainId && tokenAddress)) {
-      toast.error("Please select a chain and enter a token address");
-      return;
-    }
-    setAdding(true);
-    try {
-      await onAdd(chainId, tokenAddress);
-      setChainId(null);
-      setTokenAddress("");
-    } catch (error) {
-      toast.error(
-        error instanceof Error ? error.message : "Failed to add token"
-      );
-    } finally {
-      setAdding(false);
-    }
-  };
-
-  return (
-    <div className="space-y-3">
-      <div className="space-y-2">
-        <Label>Chain</Label>
-        <Select
-          onValueChange={(value) => setChainId(Number.parseInt(value, 10))}
-          value={chainId?.toString() ?? ""}
-        >
-          <SelectTrigger>
-            <SelectValue placeholder="Select chain" />
-          </SelectTrigger>
-          <SelectContent>
-            {chains.map((chain) => (
-              <SelectItem key={chain.chainId} value={chain.chainId.toString()}>
-                {chain.name}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </div>
-      <div className="space-y-2">
-        <Label>Token Address</Label>
-        <Input
-          disabled={adding}
-          onChange={(e) => setTokenAddress(e.target.value)}
-          placeholder="0x..."
-          value={tokenAddress}
-        />
-      </div>
-      <div className="flex gap-2">
-        <Button
-          className="flex-1"
-          disabled={adding}
-          onClick={onCancel}
-          variant="outline"
-        >
-          Cancel
-        </Button>
-        <Button
-          className="flex-1"
-          disabled={adding || !chainId || !tokenAddress}
-          onClick={handleAdd}
-        >
-          {adding ? (
-            <>
-              <Spinner className="mr-2 h-4 w-4" />
-              Adding...
-            </>
-          ) : (
-            "Add Token"
-          )}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function CreateWalletForm({
-  initialEmail,
-  onSubmit,
-}: {
-  initialEmail: string;
-  onSubmit: (email: string) => Promise<void>;
-}) {
-  const [email, setEmail] = useState(initialEmail);
-  const [creating, setCreating] = useState(false);
-
-  const handleCreate = async (): Promise<void> => {
-    if (!email) {
-      toast.error("Email is required");
-      return;
-    }
-    setCreating(true);
-    try {
-      await onSubmit(email);
-    } catch (error) {
-      toast.error(
-        error instanceof Error ? error.message : "Failed to create wallet"
-      );
-    } finally {
-      setCreating(false);
-    }
-  };
-
-  return (
-    <div className="space-y-4">
-      <p className="text-muted-foreground text-xs">
-        This wallet will be shared by all members of your organization.
-        Only admins and owners can manage it.
-      </p>
-
-      <div className="space-y-2">
-        <Label htmlFor="wallet-email">Email Address</Label>
-        <Input
-          disabled={creating}
-          id="wallet-email"
-          onChange={(e) => setEmail(e.target.value)}
-          placeholder="you@example.com"
-          type="email"
-          value={email}
-        />
-        <p className="text-muted-foreground text-xs">
-          This email will be associated with the wallet for identification
-          purposes.
-        </p>
-      </div>
-
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-1.5 text-muted-foreground text-xs">
-          <ShieldCheck className="h-3.5 w-3.5" />
-          <span>Secured by Turnkey</span>
-        </div>
-        <Button disabled={creating || !email} onClick={handleCreate}>
-          {creating ? (
-            <>
-              <Spinner className="mr-2 h-4 w-4" />
-              Creating...
-            </>
-          ) : (
-            "Create Wallet"
-          )}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-// ============================================================================
-// Section Components
-// ============================================================================
-
-function BalanceListSection({
-  balances,
-  chains,
-  isAdmin,
-  onAddToken,
-  onRefresh,
-  onRemoveToken,
-  onWithdraw,
-  refreshing,
-  showAddToken,
-  setShowAddToken,
-  supportedTokenBalances,
-  tokenBalances,
-}: {
-  balances: ChainBalance[];
-  chains: ChainData[];
-  isAdmin: boolean;
-  onAddToken: (chainId: number, tokenAddress: string) => Promise<void>;
-  onRefresh: () => void;
-  onRemoveToken: (tokenId: string, symbol: string) => void;
-  onWithdraw: (chainId: number, tokenAddress?: string) => void;
-  refreshing: boolean;
-  showAddToken: boolean;
-  setShowAddToken: (show: boolean) => void;
-  supportedTokenBalances: SupportedTokenBalance[];
-  tokenBalances: TokenBalance[];
-}) {
-  const [showTestnets, setShowTestnets] = useState(false);
-  const filteredBalances = balances.filter((b) =>
-    showTestnets ? b.isTestnet : !b.isTestnet
-  );
-
-  return (
-    <div className="space-y-3">
-      <div className="flex items-center justify-between">
-        <h3 className="font-medium text-sm">Balances</h3>
-        <div className="flex items-center gap-2">
-          <div className="flex rounded-md border">
-            <Button
-              className="h-7 rounded-r-none border-0 px-3 text-xs"
-              onClick={() => setShowTestnets(false)}
-              size="sm"
-              variant={showTestnets ? "ghost" : "default"}
-            >
-              Mainnets
-            </Button>
-            <Button
-              className="h-7 rounded-l-none border-0 px-3 text-xs"
-              onClick={() => setShowTestnets(true)}
-              size="sm"
-              variant={showTestnets ? "default" : "ghost"}
-            >
-              Testnets
-            </Button>
-          </div>
-          <Button
-            aria-label="Refresh balances"
-            data-testid="wallet-refresh-button"
-            disabled={refreshing}
-            onClick={onRefresh}
-            size="sm"
-            variant="ghost"
-          >
-            <RefreshCw
-              className={`h-4 w-4 ${refreshing ? "animate-spin" : ""}`}
-            />
-          </Button>
-        </div>
-      </div>
-
-      {balances.length === 0 ? (
-        <div className="text-muted-foreground text-sm">Loading chains...</div>
-      ) : (
-        <div className="max-h-64 space-y-2 overflow-y-auto pr-1">
-          {filteredBalances.map((balance) => (
-            <ChainBalanceItem
-              balance={balance}
-              chain={chains.find((c) => c.chainId === balance.chainId)}
-              isAdmin={isAdmin}
-              key={balance.chainId}
-              onRemoveToken={onRemoveToken}
-              onWithdraw={onWithdraw}
-              supportedTokenBalances={supportedTokenBalances}
-              tokenBalances={tokenBalances}
-            />
-          ))}
-          {filteredBalances.length === 0 && (
-            <div className="py-4 text-center text-muted-foreground text-sm">
-              No {showTestnets ? "testnet" : "mainnet"} chains available
-            </div>
-          )}
-        </div>
-      )}
-
-      {isAdmin && (
-        <div className="mt-4 border-t pt-4">
-          {showAddToken ? (
-            <AddTokenForm
-              chains={chains}
-              onAdd={onAddToken}
-              onCancel={() => setShowAddToken(false)}
-            />
-          ) : (
-            <Button
-              className="w-full"
-              onClick={() => setShowAddToken(true)}
-              variant="outline"
-            >
-              <Plus className="mr-2 h-4 w-4" />
-              Add Token
-            </Button>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function NoWalletSection({
-  isAdmin,
-  initialEmail,
-  onCreateWallet,
-}: {
-  isAdmin: boolean;
-  initialEmail: string;
-  onCreateWallet: (email: string) => Promise<void>;
-}) {
-  if (!isAdmin) {
-    return (
-      <div className="rounded-lg border bg-muted/50 p-4">
-        <p className="text-muted-foreground text-sm">
-          No wallet found for this organization. Only organization admins and
-          owners can create wallets.
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <CreateWalletForm
-      initialEmail={initialEmail}
-      onSubmit={onCreateWallet}
-    />
-  );
-}
-
-// ============================================================================
-// Main Component
-// ============================================================================
-
-type ExportStep = "idle" | "requesting" | "otp" | "verifying" | "done";
-
-function ExportPrivateKeyButton(): React.ReactElement {
-  const [open, setOpen] = useState(false);
-  const [step, setStep] = useState<ExportStep>("idle");
-  const [otpCode, setOtpCode] = useState("");
-  const [privateKey, setPrivateKey] = useState<string | null>(null);
-  const [revealed, setRevealed] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const handleOpen = async (): Promise<void> => {
-    setOpen(true);
-    setStep("requesting");
-    setError(null);
-    setOtpCode("");
-    setPrivateKey(null);
-    setRevealed(false);
-    try {
-      const res = await fetch("/api/user/wallet/export-key/request", {
-        method: "POST",
-      });
-      const data: { error?: string } = await res.json();
-
-      if (!res.ok) {
-        throw new Error(data.error ?? "Failed to send verification code");
-      }
-
-      setStep("otp");
-    } catch (err) {
-      toast.error(
-        err instanceof Error ? err.message : "Failed to send code"
-      );
-      setOpen(false);
-      setStep("idle");
-    }
-  };
-
-  const handleVerify = async (): Promise<void> => {
-    if (otpCode.length !== 6) {
-      setError("Enter the 6-digit code from your email");
-      return;
-    }
-
-    setStep("verifying");
-    setError(null);
-    try {
-      const res = await fetch("/api/user/wallet/export-key/verify", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ code: otpCode }),
-      });
-      const data: { privateKey?: string; error?: string } = await res.json();
-
-      if (!res.ok) {
-        throw new Error(data.error ?? "Verification failed");
-      }
-
-      if (!data.privateKey) {
-        throw new Error("No private key returned");
-      }
-
-      setPrivateKey(data.privateKey);
-      setRevealed(false);
-      setStep("done");
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Verification failed");
-      setStep("otp");
-    }
-  };
-
-  const handleCopy = (): void => {
-    if (!privateKey) {
-      return;
-    }
-    navigator.clipboard.writeText(privateKey);
-    toast.success("Private key copied to clipboard");
-  };
-
-  const handleClose = (): void => {
-    setOpen(false);
-    setStep("idle");
-    setOtpCode("");
-    setPrivateKey(null);
-    setRevealed(false);
-    setError(null);
-  };
-
-  return (
-    <>
-      <Button
-        className="w-full"
-        onClick={handleOpen}
-        size="sm"
-        variant="outline"
-      >
-        <KeyRound className="mr-2 h-3 w-3" />
-        Export Private Key
-      </Button>
-
-      <Dialog onOpenChange={(v) => { if (!v) { handleClose(); } }} open={open}>
-        <DialogContent className="sm:max-w-md">
-          <DialogHeader>
-            <DialogTitle>Export Private Key</DialogTitle>
-            <DialogDescription>
-              {step === "done"
-                ? "Your private key is shown below. Copy it and store it securely."
-                : "A verification code has been sent to your email."}
-            </DialogDescription>
-          </DialogHeader>
-
-          {step === "requesting" && (
-            <div className="flex items-center justify-center py-8">
-              <Spinner className="h-6 w-6" />
-            </div>
-          )}
-
-          {(step === "otp" || step === "verifying") && (
-            <div className="space-y-4 py-2">
-              <div className="space-y-2">
-                <Label htmlFor="export-otp">Verification Code</Label>
-                <Input
-                  className="font-mono text-center text-lg tracking-[0.3em]"
-                  id="export-otp"
-                  maxLength={6}
-                  onChange={(e) =>
-                    setOtpCode(e.target.value.replace(/\D/g, ""))
-                  }
-                  placeholder="000000"
-                  value={otpCode}
-                />
-                {error && (
-                  <p className="text-destructive text-sm">{error}</p>
-                )}
-              </div>
-              <div className="flex gap-2">
-                <Button
-                  className="flex-1"
-                  disabled={step === "verifying"}
-                  onClick={handleClose}
-                  variant="outline"
-                >
-                  Cancel
-                </Button>
-                <Button
-                  className="flex-1"
-                  disabled={step === "verifying" || otpCode.length !== 6}
-                  onClick={handleVerify}
-                >
-                  {step === "verifying" ? (
-                    <>
-                      <Spinner className="mr-2 h-4 w-4" />
-                      Verifying...
-                    </>
-                  ) : (
-                    "Verify & Export"
-                  )}
-                </Button>
-              </div>
-            </div>
-          )}
-
-          {step === "done" && privateKey && (
-            <div className="space-y-4 py-2">
-              <div className="rounded-lg border border-destructive/30 bg-destructive/5 p-4">
-                <div className="mb-2 flex items-center justify-between">
-                  <span className="font-medium text-destructive text-sm">
-                    Private Key
-                  </span>
-                  <div className="flex items-center gap-2">
-                    <button
-                      aria-label={
-                        revealed ? "Hide private key" : "Reveal private key"
-                      }
-                      className="text-muted-foreground hover:text-foreground"
-                      onClick={() => setRevealed(!revealed)}
-                      type="button"
-                    >
-                      {revealed ? (
-                        <EyeOff className="h-4 w-4" />
-                      ) : (
-                        <Eye className="h-4 w-4" />
-                      )}
-                    </button>
-                    <button
-                      aria-label="Copy private key"
-                      className="text-muted-foreground hover:text-foreground"
-                      onClick={handleCopy}
-                      type="button"
-                    >
-                      <Copy className="h-4 w-4" />
-                    </button>
-                  </div>
-                </div>
-                <code className="block break-all font-mono text-sm">
-                  {revealed
-                    ? privateKey
-                    : privateKey.replace(/./g, "\u2022")}
-                </code>
-              </div>
-              <Button className="w-full" onClick={handleClose}>
-                Done
-              </Button>
-            </div>
-          )}
-        </DialogContent>
-      </Dialog>
-    </>
-  );
-}
-
-function AccountDetailsSection({
-  email,
-  walletAddress,
-  isAdmin,
-  canExportKey,
-  onEmailUpdated,
-}: {
-  email: string;
-  walletAddress: string;
-  isAdmin: boolean;
-  canExportKey: boolean;
-  onEmailUpdated: () => void;
-}) {
-  const [isEditingEmail, setIsEditingEmail] = useState(false);
-  const [newEmail, setNewEmail] = useState("");
-  const [updating, setUpdating] = useState(false);
-
-  const handleUpdateEmail = async () => {
-    if (!newEmail) {
-      toast.error("Email is required");
-      return;
-    }
-
-    setUpdating(true);
-    try {
-      const response = await fetch("/api/user/wallet", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ email: newEmail }),
-      });
-
-      const data = await response.json();
-
-      if (!response.ok) {
-        throw new Error(data.error || "Failed to update email");
-      }
-
-      toast.success("Wallet email updated successfully!");
-      setIsEditingEmail(false);
-      setNewEmail("");
-      onEmailUpdated();
-    } catch (error) {
-      console.error("Failed to update wallet email:", error);
-      toast.error(
-        error instanceof Error ? error.message : "Failed to update email"
-      );
-    } finally {
-      setUpdating(false);
-    }
-  };
-
-  const startEditing = () => {
-    setNewEmail(email);
-    setIsEditingEmail(true);
-  };
-
-  const cancelEditing = () => {
-    setIsEditingEmail(false);
-    setNewEmail("");
-  };
-
-  const copyAddress = () => {
-    navigator.clipboard.writeText(toChecksumAddress(walletAddress));
-    toast.success("Address copied to clipboard");
-  };
-
-  return (
-    <div className="rounded-lg border bg-muted/50 p-4">
-      <div className="mb-2 text-muted-foreground text-sm">Account details</div>
-
-      {isEditingEmail ? (
-        <div className="space-y-2">
-          <Input
-            className="text-sm"
-            disabled={updating}
-            onChange={(e) => setNewEmail(e.target.value)}
-            placeholder="newemail@example.com"
-            type="email"
-            value={newEmail}
-          />
-          <div className="flex gap-2">
-            <Button
-              disabled={updating}
-              onClick={cancelEditing}
-              size="sm"
-              variant="outline"
-            >
-              Cancel
-            </Button>
-            <Button
-              disabled={updating || !newEmail || newEmail === email}
-              onClick={handleUpdateEmail}
-              size="sm"
-            >
-              {updating ? (
-                <>
-                  <Spinner className="mr-2 h-3 w-3" />
-                  Saving...
-                </>
-              ) : (
-                "Save"
-              )}
-            </Button>
-          </div>
-        </div>
-      ) : (
-        <div className="space-y-1">
-          <div className="flex items-center gap-2">
-            <span className="text-sm">{email}</span>
-            {isAdmin && (
-              <Button
-                className="h-5 px-1.5 text-xs"
-                onClick={startEditing}
-                size="sm"
-                variant="ghost"
-              >
-                Edit
-              </Button>
-            )}
-          </div>
-          <div className="flex items-center gap-1">
-            <code className="font-mono text-muted-foreground text-xs">
-              {truncateAddress(walletAddress)}
-            </code>
-            <button
-              aria-label="Copy wallet address"
-              className="text-muted-foreground hover:text-foreground"
-              data-testid="wallet-copy-address"
-              onClick={copyAddress}
-              type="button"
-            >
-              <Copy className="h-3 w-3" />
-            </button>
-          </div>
-        </div>
-      )}
-
-      {isAdmin && canExportKey && (
-        <div className="mt-3">
-          <ExportPrivateKeyButton />
-        </div>
-      )}
-    </div>
-  );
-}
-
-const PROVIDER_DISPLAY_ORDER: Record<"para" | "turnkey", number> = {
-  para: 0,
-  turnkey: 1,
-} as const;
-
-function WalletSwitcher({
-  wallets,
-  isAdmin,
-  switching,
-  onSelect,
-}: {
-  wallets: WalletInfo[];
-  isAdmin: boolean;
-  switching: boolean;
-  onSelect: (walletId: string) => void;
-}): React.ReactElement {
-  const ordered = [...wallets].sort(
-    (a, b) =>
-      PROVIDER_DISPLAY_ORDER[a.provider] - PROVIDER_DISPLAY_ORDER[b.provider]
-  );
-  const hasPara = ordered.some((w) => w.provider === "para");
-  return (
-    <div className="rounded-lg border bg-muted/50 p-3">
-      <div className="mb-2 text-muted-foreground text-xs">
-        Active wallet for signing
-      </div>
-      {hasPara && (
-        <div className="mb-3 rounded-md border border-amber-500/40 bg-amber-500/10 px-3 py-2 text-amber-700 text-xs dark:text-amber-400">
-          Para will be deprecated soon. Switch the active wallet to Turnkey and
-          move your assets from the Para wallet before the cutover.
-        </div>
-      )}
-      <div className="flex gap-2">
-        {ordered.map((wallet) => {
-          const label = wallet.provider === "para" ? "Para" : "Turnkey";
-          return (
-            <button
-              className={`flex-1 rounded-md border px-3 py-2 text-left text-sm transition ${
-                wallet.isActive
-                  ? "border-primary bg-primary/10"
-                  : "border-border hover:bg-background"
-              } ${!isAdmin || switching ? "cursor-not-allowed opacity-60" : ""}`}
-              disabled={!isAdmin || switching || wallet.isActive}
-              key={wallet.id}
-              onClick={() => onSelect(wallet.id)}
-              type="button"
-            >
-              <div className="flex items-center justify-between">
-                <span className="font-medium">{label}</span>
-                {wallet.isActive && (
-                  <span className="rounded bg-primary/20 px-1.5 py-0.5 text-[10px] text-primary">
-                    Active
-                  </span>
-                )}
-              </div>
-              <code className="font-mono text-muted-foreground text-xs">
-                {truncateAddress(wallet.walletAddress)}
-              </code>
-            </button>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-export function WalletOverlay({ overlayId }: WalletOverlayProps) {
+export function WalletOverlay({
+  overlayId,
+  initialTab = "balances",
+}: WalletOverlayProps) {
   const { closeAll, push } = useOverlay();
   const { data: session } = useSession();
   const { isAdmin } = useActiveMember();
@@ -1185,11 +46,21 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
   const [supportedTokenBalances, setSupportedTokenBalances] = useState<
     SupportedTokenBalance[]
   >([]);
+  const [supportedBalancesLoading, setSupportedBalancesLoading] =
+    useState(false);
   const [refreshing, setRefreshing] = useState(false);
-  const [showAddToken, setShowAddToken] = useState(false);
   const [switchingWallet, setSwitchingWallet] = useState(false);
+  // Controlled tab state so reload cycles (e.g. switching Para/Turnkey) don't
+  // reset the user's active tab back to "balances".
+  const [activeTab, setActiveTab] = useState<WalletTab>(initialTab);
 
-  const { balances, tokenBalances, fetchBalances } = useWalletBalances();
+  const {
+    balances,
+    tokenBalances,
+    loading: nativeBalancesLoading,
+    fetchBalances,
+  } = useWalletBalances();
+  const isLoadingBalances = nativeBalancesLoading || supportedBalancesLoading;
 
   const fetchChains = useCallback(async (): Promise<ChainData[]> => {
     try {
@@ -1236,13 +107,13 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
       walletAddress: string,
       chainList: ChainData[],
       tokenList: SupportedToken[]
-    ) => {
+    ): Promise<void> => {
       if (tokenList.length === 0) {
         setSupportedTokenBalances([]);
         return;
       }
 
-      // Set loading state
+      setSupportedBalancesLoading(true);
       setSupportedTokenBalances(
         tokenList.map((token) => ({
           chainId: token.chainId,
@@ -1255,13 +126,16 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
         }))
       );
 
-      // Fetch balances
-      const results = await fetchAllSupportedTokenBalances(
-        walletAddress,
-        tokenList,
-        chainList
-      );
-      setSupportedTokenBalances(results);
+      try {
+        const results = await fetchAllSupportedTokenBalances(
+          walletAddress,
+          tokenList,
+          chainList
+        );
+        setSupportedTokenBalances(results);
+      } finally {
+        setSupportedBalancesLoading(false);
+      }
     },
     []
   );
@@ -1327,7 +201,10 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
     fetchSupportedBalances,
   ]);
 
-  const handleAddToken = async (chainId: number, tokenAddress: string) => {
+  const handleAddToken = async (
+    chainId: number,
+    tokenAddress: string
+  ): Promise<void> => {
     const response = await fetch("/api/user/wallet/tokens", {
       method: "POST",
       headers: { "Content-Type": "application/json" },
@@ -1340,11 +217,13 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
     }
 
     toast.success(`Added ${data.token.symbol} to tracked tokens`);
-    setShowAddToken(false);
     await loadWallet();
   };
 
-  const handleRemoveToken = async (tokenId: string, symbol: string) => {
+  const handleRemoveToken = async (
+    tokenId: string,
+    symbol: string
+  ): Promise<void> => {
     try {
       const response = await fetch("/api/user/wallet/tokens", {
         method: "DELETE",
@@ -1358,9 +237,8 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
       }
 
       toast.success(`Removed ${symbol} from tracked tokens`);
-      await loadWallet(); // Refresh to show updated token list
+      await loadWallet();
     } catch (error) {
-      console.error("Failed to remove token:", error);
       toast.error(
         error instanceof Error ? error.message : "Failed to remove token"
       );
@@ -1413,14 +291,11 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
   const buildWithdrawableAssets = useCallback((): WithdrawableAsset[] => {
     const assets: WithdrawableAsset[] = [];
 
-    // Add native balances (skip TEMPO - it uses stablecoins, not native tokens)
     for (const balance of balances) {
-      // Skip TEMPO native balance - it uses stablecoins only
       if (isTempoChain(balance.chainId)) {
         continue;
       }
       const chain = chains.find((c) => c.chainId === balance.chainId);
-      // Skip if not a valid positive number
       const numBalance = Number.parseFloat(balance.balance);
       if (!(chain && Number.isFinite(numBalance)) || numBalance <= 0) {
         continue;
@@ -1437,9 +312,7 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
       });
     }
 
-    // Add supported token balances (stablecoins)
     for (const token of supportedTokenBalances) {
-      // Skip if not a valid positive number
       const numBalance = Number.parseFloat(token.balance);
       if (!Number.isFinite(numBalance) || numBalance <= 0) {
         continue;
@@ -1466,7 +339,11 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
   }, [balances, chains, supportedTokenBalances]);
 
   const findAssetIndex = useCallback(
-    (assets: WithdrawableAsset[], chainId: number, tokenAddress?: string) => {
+    (
+      assets: WithdrawableAsset[],
+      chainId: number,
+      tokenAddress?: string
+    ): number => {
       if (tokenAddress) {
         const idx = assets.findIndex(
           (a) => a.chainId === chainId && a.tokenAddress === tokenAddress
@@ -1482,7 +359,7 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
   );
 
   const handleWithdraw = useCallback(
-    (chainId: number, tokenAddress?: string) => {
+    (chainId: number, tokenAddress?: string): void => {
       if (!walletData?.walletAddress) {
         return;
       }
@@ -1530,40 +407,45 @@ export function WalletOverlay({ overlayId }: WalletOverlayProps) {
       )}
 
       {!walletLoading && walletData?.hasWallet && (
-        <div className="space-y-4">
-          {walletData.wallets && walletData.wallets.length > 1 && (
-            <WalletSwitcher
+        <Tabs
+          className="w-full"
+          onValueChange={(value) => setActiveTab(value as WalletTab)}
+          value={activeTab}
+        >
+          <TabsList className="w-full">
+            <TabsTrigger value="balances">Balances</TabsTrigger>
+            <TabsTrigger value="manage">Manage</TabsTrigger>
+          </TabsList>
+          <TabsContent className="mt-4" value="balances">
+            <BalancesTab
+              balances={balances}
+              chains={chains}
               isAdmin={isAdmin}
-              onSelect={handleSelectActiveWallet}
-              switching={switchingWallet}
-              wallets={walletData.wallets}
+              isLoadingBalances={isLoadingBalances}
+              onAddToken={handleAddToken}
+              onRefresh={handleRefresh}
+              onRemoveToken={handleRemoveToken}
+              onWithdraw={handleWithdraw}
+              refreshing={refreshing}
+              supportedTokenBalances={supportedTokenBalances}
+              tokenBalances={tokenBalances}
             />
-          )}
-          {walletData.email && walletData.walletAddress && (
-            <AccountDetailsSection
-              canExportKey={!!walletData.canExportKey}
-              email={walletData.email}
-              isAdmin={isAdmin}
-              onEmailUpdated={loadWallet}
-              walletAddress={walletData.walletAddress}
-            />
-          )}
-
-          <BalanceListSection
-            balances={balances}
-            chains={chains}
-            isAdmin={isAdmin}
-            onAddToken={handleAddToken}
-            onRefresh={handleRefresh}
-            onRemoveToken={handleRemoveToken}
-            onWithdraw={handleWithdraw}
-            refreshing={refreshing}
-            setShowAddToken={setShowAddToken}
-            showAddToken={showAddToken}
-            supportedTokenBalances={supportedTokenBalances}
-            tokenBalances={tokenBalances}
-          />
-        </div>
+          </TabsContent>
+          <TabsContent className="mt-4 space-y-4" value="manage">
+            {walletData.email && walletData.walletAddress && (
+              <ManageTab
+                canExportKey={!!walletData.canExportKey}
+                email={walletData.email}
+                isAdmin={isAdmin}
+                onEmailUpdated={loadWallet}
+                onSelectActiveWallet={handleSelectActiveWallet}
+                switchingWallet={switchingWallet}
+                walletAddress={walletData.walletAddress}
+                wallets={walletData.wallets}
+              />
+            )}
+          </TabsContent>
+        </Tabs>
       )}
 
       {!(walletLoading || walletData?.hasWallet) && (
